+++ src/ui/guiforcase5/EditPage.java
@@ -1,4 +1,4 @@
-package ui.incentiveui;
+package ui.guiforcase5;
 
 import javax.swing.*;
 import java.awt.*;
